Tests: demonstrate repeating subfield deletion issue
If a dataset has two `repeating_subfields` fields such that the earlier
has at least two entries, removing an entry that isn't the last breaks
the later field.

In this instance, `contact_address` is simply absent from the dataset.
In others, it could reorder the subfields.

This is because in `expand_form_composite`, the indexes dictionary
isn't scoped to the field, so here `indexes['1'] = 0` due to the deletion,
then `indexes['0'] = 1` for `contact_address`, but 1 is out of range
for the `contact_address` list.